Improve CRUD operations and introduce Student/Teacher auth
In this commit, we improved abstraction of CRUD functionalities in the base controller and renamed it to CrudControllerAbstract. We also refactored the comments for better code readability and understanding.

A new User model has been created to support Laravel's built-in authentication.

We introduced a LoginService, handling the authentication logic for both Students and Teachers, which replaces the responsibility from the previously existing User class.

In the config/auth.php file, we added new guards and providers for teachers and students.

We updated the login test, ensuring that the changes made do not affect the functionality of student login.

This set of changes makes our application more extensible and maintainable by decoupling the authentication logic from User model into two separate models - Teacher and Student. This also enhances the security of our application by providing separate authentication guards for each user type.