Refactor namespace and remove unnecessary spaces
In this commit, the namespace in LoginService has been updated from 'Services' to 'App\Services' to adhere to the standard naming conventions. Unnecessary empty lines in CrudControllerAbstract and StudentController have also been removed to improve the readability of the code. This process of cleaning and creating a clear structure makes it easier for other developers to understand the flow of the project.