Remove obsolete files and methods, improve documentation
Removed `StudentBaseRequest.php` due to it being unused and moved `TeacherBaseRequestTrait.php` into a `Traits` folder since it is a trait, not a request. Documentation comments were added for readability and understanding. Furthermore, some unnecessary fields and methods such as `teacher_id` in `Teacher.php` and `Response` import in `TeacherController.php` were removed due to lack of usage, streamlining the codebase. Naming of certain functions were updated for clarity. The token for the `LoginResource.php` was specifically fetched for usage. Furthermore, `TeacherBaseRequestTrait` was updated to align well with changes in code structure.